# docs/swagger.json
{"openapi":"3.0.0","paths":{"/api/v1/health":{"get":{"operationId":"AppController_getHealth","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/api/v1/users/find":{"get":{"operationId":"UserController_getUser","parameters":[{"name":"id","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["User"]}},"/api/v1/users":{"get":{"operationId":"UserController_getAllUsers","parameters":[],"responses":{"200":{"description":""}},"tags":["User"]}},"/api/v1/bookings":{"post":{"operationId":"BookingsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBookingDto"}}}},"responses":{"201":{"description":"Booking created successfully"},"400":{"description":"Invalid date/time parameters"},"409":{"description":"Duplicate booking slot clash"}},"summary":"Create a new booking (Public Endpoint)","tags":["Bookings"]},"get":{"operationId":"BookingsController_findAll","parameters":[{"name":"search","required":false,"in":"query","description":"Search name, email, or phone","schema":{"type":"string"}},{"name":"status","required":false,"in":"query","description":"Filter by PENDING | CONFIRMED | CANCELLED | COMPLETED","schema":{"type":"string"}},{"name":"limit","required":false,"in":"query","description":"Number of items (max 100, default 10)","schema":{"type":"number"}},{"name":"cursor","required":false,"in":"query","description":"Booking ID cursor for pagination","schema":{"type":"string"}}],"responses":{"200":{"description":"List of bookings with nextCursor pagination reference"},"401":{"description":"Unauthorized access"}},"security":[{"bearer":[]}],"summary":"Retrieve bookings with filtering and pagination","tags":["Bookings"]}},"/api/v1/bookings/{id}":{"get":{"operationId":"BookingsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Booking details"},"401":{"description":"Unauthorized access"},"404":{"description":"Booking not found"}},"security":[{"bearer":[]}],"summary":"Retrieve booking by ID","tags":["Bookings"]}},"/api/v1/bookings/{id}/status":{"patch":{"operationId":"BookingsController_updateStatus","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateBookingStatusDto"}}}},"responses":{"200":{"description":"Status updated successfully"},"400":{"description":"Illegal state transition (e.g. completing cancelled booking)"},"401":{"description":"Unauthorized access"},"404":{"description":"Booking not found"}},"security":[{"bearer":[]}],"summary":"Update status of a booking","tags":["Bookings"]}},"/api/v1/bookings/{id}/cancel":{"patch":{"operationId":"BookingsController_cancel","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Booking cancelled successfully"},"400":{"description":"Booking is already cancelled"},"401":{"description":"Unauthorized access"},"404":{"description":"Booking not found"}},"security":[{"bearer":[]}],"summary":"Cancel a booking","tags":["Bookings"]}},"/api/v1/services":{"post":{"operationId":"ServicesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateServiceDto"}}}},"responses":{"201":{"description":"Service created successfully"},"401":{"description":"Unauthorized access"}},"security":[{"bearer":[]}],"summary":"Create a new service","tags":["Services"]},"get":{"operationId":"ServicesController_findAll","parameters":[],"responses":{"200":{"description":"List of all services"},"401":{"description":"Unauthorized access"}},"security":[{"bearer":[]}],"summary":"Retrieve all services","tags":["Services"]}},"/api/v1/services/{id}":{"patch":{"operationId":"ServicesController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateServiceDto"}}}},"responses":{"200":{"description":"Service updated successfully"},"401":{"description":"Unauthorized access"},"404":{"description":"Service with the specified ID not found"}},"security":[{"bearer":[]}],"summary":"Update an existing service details","tags":["Services"]},"delete":{"operationId":"ServicesController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Service deleted successfully"},"401":{"description":"Unauthorized access"},"404":{"description":"Service not found"}},"security":[{"bearer":[]}],"summary":"Delete a service by ID","tags":["Services"]},"get":{"operationId":"ServicesController_findById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Service details"},"401":{"description":"Unauthorized access"},"404":{"description":"Service not found"}},"security":[{"bearer":[]}],"summary":"Retrieve a service by ID","tags":["Services"]}},"/api/v1/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":"User successfully registered"},"400":{"description":"Invalid input parameters"},"409":{"description":"Email address already in use"}},"summary":"Register a new user","tags":["Authentication"]}},"/api/v1/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":"User successfully logged in, returns tokens"},"401":{"description":"Invalid credentials"}},"summary":"Authenticate user credentials","tags":["Authentication"]}},"/api/v1/auth/logout":{"post":{"operationId":"AuthController_logout","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LogoutDto"}}}},"responses":{"200":{"description":"Logout successful"},"401":{"description":"Unauthorized"}},"security":[{"bearer":[]}],"summary":"Invalidate current refresh token and logout","tags":["Authentication"]}},"/api/v1/auth/refresh":{"post":{"operationId":"AuthController_refresh","parameters":[],"responses":{"200":{"description":"Tokens successfully refreshed"},"401":{"description":"Invalid refresh token"}},"security":[{"bearer":[]}],"summary":"Rotate access and refresh tokens using refresh token","tags":["Authentication"]}}},"info":{"title":"Booking Platform API","description":"API documentation for the Booking Platform REST API","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateBookingDto":{"type":"object","properties":{"serviceId":{"type":"string","example":"f28f99ce-673e-4ff4-b0eb-9b5299358f15","description":"The UUID of the service to book"},"customerName":{"type":"string","example":"John Doe","description":"The full name of the customer"},"customerEmail":{"type":"string","example":"john.doe@example.com","description":"The email address of the customer"},"customerPhone":{"type":"string","example":"+15550199","description":"The phone number of the customer"},"bookingDate":{"type":"string","example":"2026-10-15","description":"The date of the booking in YYYY-MM-DD format (must not be in the past)"},"bookingTime":{"type":"string","example":"14:30","description":"The time of the booking in 24-hour HH:MM or HH:MM:SS format"},"notes":{"type":"string","example":"Please prepare the standard consultation materials.","description":"Optional comments or requests from customer"}},"required":["serviceId","customerName","customerEmail","customerPhone","bookingDate","bookingTime"]},"UpdateBookingStatusDto":{"type":"object","properties":{"status":{"type":"string","enum":["PENDING","CONFIRMED","CANCELLED","COMPLETED"],"example":"CONFIRMED","description":"The new status of the booking"}},"required":["status"]},"CreateServiceDto":{"type":"object","properties":{"title":{"type":"string","example":"Massage Therapy","description":"The title of the service"},"description":{"type":"string","example":"Full body relaxing massage with aromatherapy oils","description":"Description of the service details"},"duration":{"type":"number","example":60,"description":"Duration of the service in minutes (minimum 1)"},"price":{"type":"number","example":85.5,"description":"Price of the service in standard currency"}},"required":["title","duration","price"]},"UpdateServiceDto":{"type":"object","properties":{"title":{"type":"string","example":"Massage Therapy","description":"The title of the service"},"description":{"type":"string","example":"Full body relaxing massage with aromatherapy oils","description":"Description of the service details"},"duration":{"type":"number","example":60,"description":"Duration of the service in minutes (minimum 1)"},"price":{"type":"number","example":85.5,"description":"Price of the service in standard currency"},"isActive":{"type":"boolean","example":true,"description":"Flag indicating if the service is active and bookable"}}},"RegisterDto":{"type":"object","properties":{"email":{"type":"string","example":"merchant@example.com","description":"The email address of the merchant"},"password":{"type":"string","example":"Password123!","description":"The password of the merchant (min 6 chars)"},"name":{"type":"string","example":"Spa Salon","description":"The name of the merchant/user"}},"required":["email","password","name"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string","example":"merchant@example.com","description":"The email address of the merchant"},"password":{"type":"string","example":"Password123!","description":"The password of the merchant"}},"required":["email","password"]},"LogoutDto":{"type":"object","properties":{"refreshToken":{"type":"string","example":"eyJhbGciOi...","description":"Optional refresh token to invalidate on logout"}}}}}}