# src/services/dto/create-service.dto.ts
import { IsNotEmpty, IsString, IsInt, Min, IsNumber, IsOptional, IsPositive } from 'class-validator';
import { ApiProperty } from '@nestjs/swagger';

export class CreateServiceDto {
    @ApiProperty({ example: 'Massage Therapy', description: 'The title of the service' })
    @IsString()
    @IsNotEmpty({ message: 'Title is required' })
    title: string;

    @ApiProperty({ example: 'Full body relaxing massage with aromatherapy oils', description: 'Description of the service details', required: false })
    @IsString()
    @IsOptional()
    description?: string;

    @ApiProperty({ example: 60, description: 'Duration of the service in minutes (minimum 1)' })
    @IsInt()
    @Min(1, { message: 'Duration must be at least 1 minute' })
    duration: number;

    @ApiProperty({ example: 85.50, description: 'Price of the service in standard currency' })
    @IsNumber()
    @IsPositive({ message: 'Price must be a positive number' })
    price: number;
}